added colors to players names based on their team
