+++ src/main/java/net/tiagofar78/prisonescape/game/Guard.java
@@ -10,9 +10,7 @@ import net.tiagofar78.prisonescape.managers.ConfigManager;
 import net.tiagofar78.prisonescape.managers.GameManager;
 import net.tiagofar78.prisonescape.managers.MessageLanguageManager;
 
-import org.bukkit.ChatColor;
 import org.bukkit.GameMode;
-import org.bukkit.scoreboard.Team;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -30,15 +28,10 @@ public class Guard extends PrisonEscapePlayer {
 
     private PrisonEscapeLocation _locationBeforeWatchingCameras = null;
 
-    private ScoreboardData _scoreboardData;
-
     public Guard(String name) {
         super(name);
 
         _balance = ConfigManager.getInstance().getStartingBalance();
-
-        _scoreboardData = createScoreboardData();
-        setScoreboard(_scoreboardData.getScoreboard());
     }
 
     @Override
@@ -124,43 +117,18 @@ public class Guard extends PrisonEscapePlayer {
 //  #              Scoreboard              #
 //  ########################################
 
+    @Override
     public ScoreboardData createScoreboardData() {
         MessageLanguageManager messages = MessageLanguageManager.getInstanceByPlayer(getName());
 
-        ScoreboardData sbData = new ScoreboardData();
+        ScoreboardData sbData = super.createScoreboardData();
 
         String balanceLine = messages.getGuardSideBarBalanceLine(getBalance());
         String lastLine = messages.getSideBarLastLine();
         List<String> baseSideBar = buildBaseSideBar(balanceLine, lastLine);
         sbData.createSideBar(messages.getScoreboardDisplayName(), baseSideBar);
 
-        String guardsTeamName = GameManager.getGame().getGuardsTeam().getName();
-        registerTeam(sbData, guardsTeamName, ChatColor.BLUE);
-
-        String prisionersTeamName = GameManager.getGame().getPrisionerTeam().getName();
-        registerTeam(sbData, prisionersTeamName, ChatColor.GOLD);
-
         return sbData;
-    }
-
-    private void registerTeam(ScoreboardData sbData, String teamName, ChatColor color) {
-        Team sbTeam = sbData.registerTeam(teamName);
-        sbTeam.setColor(color);
-    }
-
-    public void updateScoreaboardTeams() {
-        PrisonEscapeTeam<Guard> guardsTeam = GameManager.getGame().getGuardsTeam();
-        addScoreboardTeamMembers(guardsTeam);
-
-        PrisonEscapeTeam<Prisioner> prisionersTeam = GameManager.getGame().getPrisionerTeam();
-        addScoreboardTeamMembers(prisionersTeam);
-    }
-
-    private void addScoreboardTeamMembers(PrisonEscapeTeam<? extends PrisonEscapePlayer> team) {
-        Team sbTeam = _scoreboardData.getScoreboard().getTeam(team.getName());
-        for (PrisonEscapePlayer player : team.getMembers()) {
-            sbTeam.addEntry(player.getName());
-        }
     }
 
     private List<String> buildBaseSideBar(String balanceLine, String lastLine) {
@@ -181,27 +149,30 @@ public class Guard extends PrisonEscapePlayer {
     private void updateBalanceLine() {
         MessageLanguageManager messages = MessageLanguageManager.getInstanceByPlayer(getName());
         String balanceLine = messages.getGuardSideBarBalanceLine(getBalance());
-        _scoreboardData.updateLine(BALANCE_LINE_INDEX, balanceLine);
+        getScoreboardData().updateLine(BALANCE_LINE_INDEX, balanceLine);
     }
 
     public void addSoundDetectorLine(int value) {
         int soundDetectorsAmount = GameManager.getGame().getPrison().countSoundDetectors();
         if (soundDetectorsAmount == 0) {
             String emptyLine = "§a";
-            _scoreboardData.addLine(SOUND_DETECTORS_FIRST_LINE_INDEX, emptyLine);
+            getScoreboardData().addLine(SOUND_DETECTORS_FIRST_LINE_INDEX, emptyLine);
 
             MessageLanguageManager messages = MessageLanguageManager.getInstanceByPlayer(getName());
             String line = messages.getGuardSideBarSoundDetectorLine();
-            _scoreboardData.addLine(SOUND_DETECTORS_FIRST_LINE_INDEX, line);
+            getScoreboardData().addLine(SOUND_DETECTORS_FIRST_LINE_INDEX, line);
         }
 
         String soundDetectorValueLine = createSoundDetectorValueLine(soundDetectorsAmount + 1, value);
-        _scoreboardData.addLine(SOUND_DETECTORS_FIRST_LINE_INDEX + 1 + soundDetectorsAmount, soundDetectorValueLine);
+        getScoreboardData().addLine(
+                SOUND_DETECTORS_FIRST_LINE_INDEX + 1 + soundDetectorsAmount,
+                soundDetectorValueLine
+        );
     }
 
     public void updateSoundDetectorValue(int index, int value) {
         String soundDetectorValueLine = createSoundDetectorValueLine(index + 1, value);
-        _scoreboardData.updateLine(SOUND_DETECTORS_FIRST_LINE_INDEX + 1 + index, soundDetectorValueLine);
+        getScoreboardData().updateLine(SOUND_DETECTORS_FIRST_LINE_INDEX + 1 + index, soundDetectorValueLine);
     }
 
     private String createSoundDetectorValueLine(int index, int value) {
